Step.02-Solution-Populate-the-GameViewModel
